Try to use packrattle to parse layout
This is as far as it gets.
Packrattle caches the results of parsers in order to keep the complexity
down.  Every position of the input is only visited once by any one
parser.  This means that it is not possible to use the same parsers for
different "levels" of the hierarchy.